# Remove dependency on stdlib for ParseIPPort, make faster.