Makes long strings parse a little faster, because unlike net.SplitHostPort, we
can entirely ignore the contents of the ip and port portions, and let our subsequent
parser deal with errors there.

```
ParseIPPort/v4-8             56.5ns ± 3%    57.4ns ± 3%     ~     (p=0.421 n=5+5)
ParseIPPort/v6-8              143ns ± 2%     117ns ± 4%  -18.44%  (p=0.008 n=5+5)
ParseIPPort/v6_ellipsis-8     112ns ± 5%     100ns ± 1%  -10.68%  (p=0.008 n=5+5)
ParseIPPort/v6_v4-8           143ns ± 5%     139ns ± 1%     ~     (p=0.175 n=5+5)
ParseIPPort/v6_zone-8         272ns ± 1%     256ns ± 1%   -5.96%  (p=0.008 n=5+5)
```

Signed-off-by: David Anderson <dave@natulte.net>