@danderson 2020-12-30
Note this adds a test case that diverges from the behavior of net: in netaddr.ParseIPPort, enclosing IPv4 addresses in brackets is not allowed, so `[1.2.3.4]:1234` will result in a parse error. Prior to this change we allowed this syntax because net.SplitHostPort doesn't care. That said, I _think_ brackets-with-IPv4 is supposed to not parse.