@danderson 2020-12-30
Source: https://tools.ietf.org/html/rfc3986#section-3.2.2 . URI syntax says that square brackets are only valid for IPv6 or future protocols, and IPv4 must be a naked dotted-decimal. So, I think this new behavior (and deviation from stdlib) is correct?