@josharian 2020-12-31
> So, I think this new behavior (and deviation from stdlib) is correct?

If we're deviating from package net and it's a bug fix, we should probably file an upstream bug so they can fix too.